+++ src/main/java/com/jsoncomparator/controller/JsonController.java
@@ -1,5 +1,6 @@
-package com.jsoncomparator;
+package com.jsoncomparator.controller;
 
+import com.jsoncomparator.gateway.AWSGateway;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
